# public/openapi/apis.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "isrvd API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API 服务"
    }
  ],
  "paths": {
...
    "/cron/jobs": {
      "get": {
        "operationId": "get_cron_jobs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "列出计划任务",
        "tags": [
          "cron"
        ]
      },
      "post": {
        "operationId": "post_cron_jobs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/cron.JobUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "创建计划任务",
        "tags": [
          "cron"
        ]
      }
    },
    "/cron/jobs/{id}": {
      "delete": {
        "operationId": "delete_cron_jobs_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "删除计划任务",
        "tags": [
          "cron"
        ]
      },
      "patch": {
        "operationId": "patch_cron_jobs_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/cronJobEnableReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "启用或禁用任务",
        "tags": [
          "cron"
        ]
      },
      "put": {
        "operationId": "put_cron_jobs_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/cron.JobUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "更新计划任务",
        "tags": [
          "cron"
        ]
      }
    },
    "/cron/jobs/{id}/logs": {
      "get": {
        "operationId": "get_cron_jobs_id_logs",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "查询任务执行历史",
        "tags": [
          "cron"
        ]
      }
    },
    "/cron/jobs/{id}/run": {
      "post": {
        "operationId": "post_cron_jobs_id_run",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "立即执行任务",
        "tags": [
          "cron"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIResponse": {
        "properties": {
          "message": {
            "type": "string"
          },
          "payload": {},
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "cron.JobUpsertRequest": {
        "properties": {
          "container": {
            "description": "DOCKER_CTR：目标容器名",
            "type": "string"
          },
          "content": {
            "description": "脚本内容",
            "type": "string"
          },
          "description": {
            "description": "任务描述",
            "type": "string"
          },
          "enabled": {
            "description": "是否启用",
            "type": "boolean"
          },
          "image": {
            "description": "DOCKER_TMP：镜像名",
            "type": "string"
          },
          "name": {
            "description": "任务名称",
            "type": "string"
          },
          "schedule": {
            "description": "cron 表达式",
            "type": "string"
          },
          "timeout": {
            "description": "超时（秒）",
            "type": "integer"
          },
          "type": {
            "description": "脚本类型",
            "type": "string"
          },
          "volumes": {
            "description": "DOCKER_TMP：额外挂载",
            "type": "string"
          },
          "workDir": {
            "description": "工作目录",
            "type": "string"
          }
        },
        "required": [
          "name",
          "schedule",
          "type",
          "content"
        ],
        "type": "object"
      },
      "cronJobEnableReq": {
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  }
}